// src/app/shared/interfaces/estudiantes.ts
export interface Estudiantes {
    id: number;
    nombre: string;
    apellido: string;
    curso: string;
    nota: number;
    usuario: string;
  }
  
  export interface EstudiantesLista {
    id: number;
    nombre: string;
    apellido: string;
    edad:number;
    correo: string;
    telefono: number;
  
  
  }